* commit '07c8dc4e6d3d4e841b132fb40b6e9e237d28d56c': (98 commits)
  [CodeFactor] Apply fixes to commit 5c2e1f4
  Fixed: codefactor issues Update: readme
  [CodeFactor] Apply fixes to commit 5689a8e
  - Fixed: get changelog issue for some premium plugins - Fixed: issue for file uploader - Fixed: js error for child pages - Updated: hide clone tab for disconnect sites
  Update README.md
  Update README.md
  Update readme
  Updated: readme and version number
  [CodeFactor] Apply fixes to commit 5a9cd05
  Added: plugins/themes searching options
  Fixed: db updater issue
  Fixed: codefactor issues
  Fixed: codefactor issues
  Updated: APIs saving Fixed: codefactor issues
  Update: phpcs rules
  Updated: phpcs rules
  Update: phpcs rules
  Update: phpcs rules
  Fixed: codefactor issues
  [CodeFactor] Apply fixes to commit 1fcfa09
  ...

# Conflicts:
#	.gitignore

+++ class/class-mainwp-child-wp-rocket.php
@@ -330,7 +330,7 @@ class MainWP_Child_WP_Rocket {
 	 * @uses \MainWP\Child\MainWP_Helper::is_updates_screen()
 	 */
 	public function remove_update_nag( $value ) {
-		if ( isset( $_POST['mainwpsignature'] ) ) {
+		if ( MainWP_Helper::is_dashboard_request() ) {
 			return $value;
 		}
 
@@ -436,8 +436,8 @@ class MainWP_Child_WP_Rocket {
 
 		$information = array();
 
-		if ( isset( $_POST['mwp_action'] ) ) {
-			$mwp_action = ! empty( $_POST['mwp_action'] ) ? sanitize_text_field( wp_unslash( $_POST['mwp_action'] ) ) : '';
+		$mwp_action = MainWP_System::instance()->validate_params( 'mwp_action' );
+		if ( ! empty( $mwp_action ) ) {
 			try {
 				switch ( $mwp_action ) {
 					case 'set_showhide':
@@ -489,7 +489,7 @@ class MainWP_Child_WP_Rocket {
 	 * @uses    \MainWP\Child\MainWP_Helper::update_option()
 	 */
 	public function set_showhide() {
-		$hide = isset( $_POST['showhide'] ) && ( 'hide' === $_POST['showhide'] ) ? 'hide' : '';
+		$hide = MainWP_System::instance()->validate_params( 'showhide' );
 		MainWP_Helper::update_option( 'mainwp_wprocket_hide_plugin', $hide );
 		$information['result'] = 'SUCCESS';
 
@@ -579,19 +579,26 @@ class MainWP_Child_WP_Rocket {
 	 */
 	public function preload_cache() {
 		MainWP_Helper::instance()->check_functions( array( 'run_rocket_sitemap_preload', 'run_rocket_bot' ) );
-		MainWP_Helper::instance()->check_classes_exists( '\WP_Rocket\Preload\Full_Process' );
+		$existed = MainWP_Helper::instance()->check_classes_exists( '\WP_Rocket\Engine\Preload\FullProcess', true );
+		// compatible.
+		if ( true === $existed ) {
+			$preload_process = new \WP_Rocket\Engine\Preload\FullProcess();
+			MainWP_Helper::instance()->check_methods( $preload_process, array( 'is_process_running' ) );
 
-		$preload_process = new \WP_Rocket\Preload\Full_Process();
-		MainWP_Helper::instance()->check_methods( $preload_process, array( 'is_process_running' ) );
+			if ( $preload_process->is_process_running() ) {
+				return array( 'result' => 'RUNNING' );
+			}
 
-		if ( $preload_process->is_process_running() ) {
-			return array( 'result' => 'RUNNING' );
+			delete_transient( 'rocket_preload_errors' );
+			run_rocket_bot( 'cache-preload', '' );
+			run_rocket_sitemap_preload();
+			return array( 'result' => 'SUCCESS' );
+		} else {
+			// Preload cache.
+			run_rocket_bot();
+			run_rocket_sitemap_preload();
+			return array( 'result' => 'SUCCESS' );
 		}
-
-		delete_transient( 'rocket_preload_errors' );
-		run_rocket_bot( 'cache-preload', '' );
-		run_rocket_sitemap_preload();
-		return array( 'result' => 'SUCCESS' );
 	}
 
 	/**
@@ -670,6 +677,7 @@ class MainWP_Child_WP_Rocket {
 	 * @return array Action result.
 	 */
 	public function save_settings() {
+		// phpcs:disable WordPress.Security.NonceVerification
 		$options = isset( $_POST['settings'] ) ? json_decode( base64_decode( wp_unslash( $_POST['settings'] ) ), true ) : ''; // phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions -- base64_encode function is used for http encode compatible..
 		if ( ! is_array( $options ) || empty( $options ) ) {
 			return array( 'error' => 'INVALID_OPTIONS' );
@@ -689,7 +697,7 @@ class MainWP_Child_WP_Rocket {
 		if ( isset( $_POST['do_database_optimization'] ) && ! empty( $_POST['do_database_optimization'] ) ) {
 			$this->optimize_database();
 		}
-
+		// phpcs:enable WordPress.Security.NonceVerification
 		return array( 'result' => 'SUCCESS' );
 	}
